Split infer info
